% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPASE.r
\name{createPhasedBlockList}
\alias{createPhasedBlockList}
\title{Create a filtered RPASE phased block list from VCF data and an annotation table}
\usage{
createPhasedBlockList(phased_vcf_data, annotation, min_phased_sites = 2,
  min_coverage = 10, heter = 2)
}
\arguments{
\item{phased_vcf_data}{A data.frame with columns ordered according to the VCF standard containing information about phased SNPs for a single sample}

\item{annotation}{A data.frame containing annotations with each line describing a study unit, typically a gene, ordered as a four-column BED format}

\item{min_phased_sites}{minimum number of phased SNPs to include within a phased block}

\item{min_coverage}{minimum coverage for each phased SNP included within a phased block}

\item{heter}{minimum minor-allele coverage for a phased SNP to be included within a phased block as heterzugous SNP}
}
\value{
A list of data.frames, each describing a phased block, typically a gene. Each dataframe contains 6 columns: chromosome, position, index of phasing, gene, AD1 (allele depth of first allele), and AD2 (allele depth of second allele).
}
\description{
Create a filtered RPASE phased block list from VCF data and an annotation table
}
\details{
phased_vcf_data describes a single sample, and should thus have 10 columns corresponding to VCF columns CHROM, POS, ID, REF, ALT, QUAL, FILTER, INFO, FORMAT, and the sample name. annotation contains four columns corresponding to a four-column BED format: chromosome, gene start (0-based), gene end, and gene name.  The extent of phased variants within phased_vcf_data can restrict the extend of phased blocks returned, such that more than one phased block may be returned per annotation. Phased blocks that do not meet the thresholds of number of phased SNPs, coverage or heterzygosity as determined from input arguments are filtered out.
}
\examples{
data(example_phased_vcf)
example_phased_vcf
data(example_annotation)
example_annotation
createPhasedBlockList(example_phased_vcf, example_annotation)
}
