% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPASE.r
\name{createUnfilteredPhasedBlockList}
\alias{createUnfilteredPhasedBlockList}
\title{Create an unfiltered RPASE phased block list from VCF data and an annotation table}
\usage{
createUnfilteredPhasedBlockList(phased_vcf_data, annotation)
}
\arguments{
\item{phased_vcf_data}{A data.frame with columns ordered according to the VCF standard containing information about phased SNPs}

\item{annotation}{A data.frame containing gene annotations ordered as a four-column BED format}
}
\value{
A list of data.frames, each describing a phased block as determined from inputs
}
\description{
Create an unfiltered RPASE phased block list from VCF data and an annotation table
}
\details{
For further details, see the documentation for \code{\link{createPhasedBlockList}}
}
