% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPASE.r
\name{dbinorm32}
\alias{dbinorm32}
\title{Computing binomial normal marginal density using 32-points gaussian quadrature.}
\source{
Pawitan Y. In all likelihood: statistical modelling and inference using likelihood[M]. Oxford University Press, 2001.
}
\usage{
dbinorm32(x, n, p, sig)
}
\arguments{
\item{x}{a quantile}

\item{n}{number of trials}

\item{p}{probability of success on each trial}

\item{sig}{overdispersion}
}
\value{
The marginal probability density of binomial normal distribution
}
\description{
Computing binomial normal marginal density using 32-points gaussian quadrature.
}
