% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.r
\docType{data}
\name{example_annotation}
\alias{example_annotation}
\title{Example input annotation data}
\format{A data.frame containing gene annotations ordered as a four-column BED format}
\usage{
example_annotation
}
\description{
For further details, see runMakeGeneList_and_Filter
}
\keyword{RPASE}
\keyword{data}
\keyword{example}
