% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.r
\docType{data}
\name{example_phased_block_list}
\alias{example_phased_block_list}
\title{Example input of a list of phased block}
\format{A list of data.frame of each containing a phased block.In each data.frame, it should contain 6 colunms: 1)chromosome, 2)position, 3)index(of phasing), 4)gene, 5)AD1(counts from one allele), and 6)AD2(counts from another allele).}
\usage{
example_phased_block_list
}
\description{
Instead of generating a list of phased block using functions runMakeGeneList_and_Filter, one can also alternatively provide a such list with the right formart as shown here.
}
\keyword{RPASE}
\keyword{data}
\keyword{example}
