% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.r
\docType{data}
\name{example_phased_vcf}
\alias{example_phased_vcf}
\title{Example input vcf data.}
\format{A data.frame with columns ordered according to the VCF standard containing information about phased SNPs}
\usage{
example_phased_vcf
}
\description{
For further details, see runMakeGeneList_and_Filter
}
\keyword{RPASE}
\keyword{data}
\keyword{example}
