% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPASE.r
\name{RPASE}
\alias{RPASE}
\title{Calculate an exact RPASE p-value for one phased block}
\usage{
RPASE(phased_block, phi, quiet = FALSE)
}
\arguments{
\item{phased_block}{A single data.frame describing a phased block, typically a gene, as described in createPhasedBlockList}

\item{phi}{The overdispersion estimation}

\item{quiet}{A logical, controls if one should print out the results while calculating}
}
\value{
An exact p-value for one phased block, typically a gene
}
\description{
Calculate an exact RPASE p-value for one phased block
}
\details{
This function implements the RPASE method for Read-backed Phasing detection of Allele-Specific Expression.  For further methodological details, see the corresponding paper (Wang et al. 2017).  Rarely, this method can require considerable time to complete when coverage is deep and variance in coverage between sites within a block (gene) is high.  In our experience this occurs in less than 0.05% of cases, and occurs most often with deeper coverage and higher SNP densities, both of which are more likely in repetitive regions.
}
\examples{
data(example_phased_block_list)
example_phased_block_list
RPASE(example_phased_block_list[[1]], phi=0.3)
}
