% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPASE.r
\name{runRPASE}
\alias{runRPASE}
\title{Calculate exact RPASE p-values for each element of a list of phased blocks}
\usage{
runRPASE(phased_block_list, phi = phi.estimate(phased_block_list),
  quiet = FALSE)
}
\arguments{
\item{phased_block_list}{A list of data.frames each describing a phased block, typically a gene. This data structure is created by createPhasedBlockList}

\item{phi}{The overdispersion estimate; by default, it is estimated from the phased_block_list}

\item{quiet}{A logical, controls if one should print out the results while calculating}
}
\value{
A vector of exact p-values
}
\description{
Calculate exact RPASE p-values for each element of a list of phased blocks
}
\details{
This function implements the RPASE method for detecting allele-specific expression using read-backed phasing by evaluating each phased block in phased_block_list in sequence.  Alternatively, this procedure could be parallelized by using the RPASE function itself together with functions from the parallel package or another such package.
}
\examples{
data(example_phased_block_list)
runRPASE(example_phased_block_list, phi=0.3)
}
